requirements.txt
# Django framework
Django>=5.0,<5.2

# Django rest framework
djangorestframework
djangorestframework-simplejwt

# CORS security
django-cors-headers

# YAML parsing
pyyaml

# Managing environment
python-decouple

# Processing requests
requests

# SSL certificates on local machine
certifi

# External search API
algoliasearch-django>=3.0,<4.0